Moved ammo to own file
